+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to the "vscode-redmine" extension will be documented in this file.
 
 ## [Unreleased]
+- Added possibility to change issue status
 
 ## 0.0.2 - 28.01.2018
 - Added repository URL to `package.json`, so VSCode Marketplace see `README.md` and other files